fix -O3 for sysvchecksum

+++ src/hash/sysvchecksum.h
@@ -7,8 +7,8 @@ struct sysvchecksum_hash {
 
 struct sysvchecksum_hash sysvchecksum_init();
 void sysvchecksum_update(uint8_t*, size_t, struct sysvchecksum_hash*);
-uint32_t sysvchecksum_final(struct sysvchecksum_hash*);
-uint32_t sysvchecksum(uint8_t*, size_t);
+uint16_t sysvchecksum_final(struct sysvchecksum_hash*);
+uint16_t sysvchecksum(uint8_t*, size_t);
 
 int l_sysvchecksum(lua_State*);
 int l_sysvchecksum_init(lua_State*);
